Node gRPC API services refactor to use abstract effect type
