+++ node/src/main/scala/coop/rchain/node/api/ReplGrpcService.scala
@@ -1,21 +1,22 @@
 package coop.rchain.node.api
 
 import cats.effect.Sync
-import cats.implicits._
-import coop.rchain.catscontrib.Catscontrib._
-import coop.rchain.catscontrib._
+import cats.syntax.all._
 import coop.rchain.models.Par
+import coop.rchain.monix.Monixable
 import coop.rchain.node.model.repl._
 import coop.rchain.rholang.interpreter.Interpreter._
 import coop.rchain.rholang.interpreter.compiler.ParBuilder
 import coop.rchain.rholang.interpreter.errors.InterpreterError
 import coop.rchain.rholang.interpreter.storage.StoragePrinter
 import coop.rchain.rholang.interpreter.{Runtime, _}
+import coop.rchain.shared.syntax._
 import monix.eval.Task
 import monix.execution.Scheduler
 
 object ReplGrpcService {
-  def instance[F[_]: Sync: Taskable](runtime: Runtime[F], worker: Scheduler): ReplGrpcMonix.Repl =
+
+  def apply[F[_]: Monixable: Sync](runtime: Runtime[F], worker: Scheduler): ReplGrpcMonix.Repl =
     new ReplGrpcMonix.Repl {
       def exec(source: String, printUnmatchedSendsOnly: Boolean = false): F[ReplResponse] =
         Sync[F]
@@ -55,7 +56,7 @@ object ReplGrpcService {
           .map(ReplResponse(_))
 
       private def defer[A](task: F[A]): Task[A] =
-        Task.defer(task.toTask).executeOn(worker)
+        task.toTask.executeOn(worker)
 
       def run(request: CmdRequest): Task[ReplResponse] =
         defer(exec(request.line))
